+++ blockfrost-api/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Version [next](https://github.com/blockfrost/blockfrost-haskell/compare/api-0.12.1.0...master) (2024-MM-DD)
 
+* Additions
+  * `/blocks/:hash_or_number/txs/cbor`
+    and `/blocks/latest/txs/cbor` endpoints with `TxHashesCBOR` data type [#73](https://github.com/blockfrost/blockfrost-haskell/pull/73)
+
 # Version [0.12.1.0](https://github.com/blockfrost/blockfrost-haskell/compare/api-0.12.0.0...api-0.12.1.0) (2024-09-25)
 
 * Fix return type of `/txs/:hash/cbor` endpoint from `[TransactionCBOR]` to `TransactionCBOR`.